Unix socket directory update during upgrade
This update set the unix_socket_directories parameter
in the PG configuration to /tmp.

This is done to ensure any existing references to the /crunchyadm
directory are no longer set in the PG configuration when a cluster
is upgraded.

As this value is stored in the DCS according to the key/value pairs
defined in the PG configmap, it is corrected by updating the
<clusterName>-pgha-config ConfigMap to no longer reference the
removed directory.

Similarly, any existing references to scripts in the previous
/opt/cpm paths stored in existing configmaps will be updated and
synchronized with the DCS to the current /opt/crunchy paths currently
defined in the associated containers.