Clean Stanza Create pgtask & Job After Init
When a PostgreSQL cluster is initialized for the first time, the stanza
create pgtask and the Kubernetes Job associated with it are now
automatically deleted.  This ensures that subsequent update events in
the Job controller for a stanza-create Job do not trigger additional
"initial backups" Jobs for the cluster.

Issue: [ch10277]
Issue: #2106