Check nested struct initialization during CM upgrade
During the upgrade of an existing pgcluster, the existing
<clustername>-pgha-config configmap is used to populate
the struct object so that key values are upgraded, as needed.

In certain cases where the configmap exists but was is not
initialized as expected, the exist function can hit a runtime
panic when referencing uninitialized structs or maps.

This update checks that these structs and maps are not nil
before attempting to set any values.