Ensure proper image tag values are authoritative
During both pgcluster create and upgrade, the --ccp-image-tag
flag can be used to specify an image tag that overrides the
value set in the PostgreSQL Operator default configuration.
In certain cases, this value is not authoritative because the
underlying ccpimagetag CRD value is not honored as expected.

This update standardizes the behavior throughout the Operator so
that the CLI flag and subsequent CRD value is authoritative both
during initial cluster creation/upgrade and also used during
subsequent opertions, such as backups and restores.

This update is designed to conform to the standard ccpimagetag
pattern. For example, the tag

centos8-12.5-4.6.0

will correspond to the image OS version, PostgreSQL version and
Crunchy PostgreSQL Operator version. Similarly, a PostGIS enabled
image would be expected to have a tag similar to

centos8-12.5-3.0-4.6.0

where the additional value, 3.0, corresponds to the PostGIS version.
This is particularly important to note when using PostGIS enabled
clusters, as the Operator will automatically handle the removal of
the '3.0' value when creating sidecars or companion pods, such as
pgBadger or pgBouncer.

Finally, the tag validation used during the Operator's automated
upgrade process has been adjusted to properly perform tag validation
based on the provide image tag when the --ccp-image-tag flag is used.