Update restore e2e test
Restores now keep the original pvc and perform a delta restore instead of
re-creating the pvc and running a full restore. This test was checking to
confirm that the pvc was re-created as part of the restore process. Since the
original pvc is not re-created, we should now check that the creation time for
the cluster pvc pre-restore and post-restore is the same.