Existence check for primary Deployment before scaling
During the cluster initialization process, there is a step where
the Deployment object representing the primary is explicitly
scaled from 0 to 1 Pods. However, there is a case that may
trigger the Deployment scaling before the Deployment is created.
As such, we can add a guard to check that the Deployment is
actually created before proceeding with the scaling operation.

This patch also modifies a debug line around the cluster
Deployment creation, as the debug output came after the
point at which the code may have exited due to error.