Use a liveness probe to stop an unresponsive Patroni
If the Patroni API stops responding, there's a good chance it has
stopped managing PostgreSQL as well. In the worst case, this can lead to
multiple instances in the cluster committing transactions, a.k.a. a
split-brain state.

For this very reason, Patroni automatically activates a system watchdog
when it is available and the instance is leader. Kubernetes liveness
probe serves a similar purpose, though it is distinctly less flexible
than a watchdog device. One major drawback is that the probe is
immutable; changing it requires the Pod to be redeployed.

Issue: [ch10192]