Automatically detect proper setting for "disable FSGroup"
In particular, we want to disable explicitly setting the FSGroup
when running in OpenShift, i.e. when using the "restricted" SCC
(which is the default).

This adds detection to see if we are within an OpenShift cluster
and, if so, will automatically disable the FSGroup.

This also removes the default explicit setting of the disable
FSGroup feature from the installer. However, one can still set
this option should they so desire.

Issue: [ch10571]