Revert changes from 05742298
05742298 was designed to workaround an issue with the Docker
packaging from Red Hat that introduced a bug which prevented the
containers from properly loading. Given we are in a single-user
environment that only uses ssh public key authentication for
performing pgBackRest operations, we can set "UsePAM" to "no".

This allows for one to run the containers with a security context
of "allowPrivilegeEscalation" set to "false".

Issue: [ch10588]