Create Postgres user connection secret
This commit creates a Kubernetes secret that contains
the core information needed to connect to a postgrescluster's
primary database. The secret contains the username, password,
database name and a connection string in the format of
postgresql://user:secret@localhost:port/mydb
where 'user' is the username, 'secret' is the password, 'localhost'
is the hostname, 'port' is the port and 'mydb' is the database name
See
https://www.postgresql.org/docs/current/libpq-connect.html#LIBPQ-CONNSTRING
for more information.