Use Real Volumes for PGDATA
Real volumes are now used for the PGDATA directories corresponding to
the various instances within a PostgreSQL cluster.  Specifically, a
PersistentVolumeClaim spec must now be specified for each instance
defined within a PostgresCluster spec. This spec is used to create and
bind the PersistentVolumeClaim(s) that will be utilized to mount PGDATA
volumes for each instance within the PostgreSQL cluster.

With this change, the PGDATA directory itself has been moved from
'/tmp' to '/pgdata', with the configuration files for all applicable
applications (i.e. Patroni and pgBackRest) updated accordingly.
Additionally, a subdirectory is now defined underneath the '/pgdata'
directory according to the specific PostgreSQL version being run.  And
to capture the proper PostgreSQL version, a field has been added to the
PostgresCluster spec that must be provided by the user.

Additionally, in support of easier image tag configuration (e.g. the
ability to use re-tagged containers or different container builds), the
hard-coded 'postgres-ha' container name has been removed from the
database container definition, and is now instead provided by the user
via an 'image' field within the PostgresCluster spec.  And finally,
supplemental groups '65534' are now added by default to the PostgreSQL
and pgBackRest containers.