Adds PGDATA Init Container
Adds a PGDATA initialization container to all PostgresCluster Pods,
which is responsible for creating the PGDATA directory if it doesn't
exist, and then setting the proper permissions for that directory in
accordance with PostgreSQL requirements. This ensures that the PGDATA
directory permissions are properly configured after being modified
by Kubernetes when setting fsGroups.