Enhance default TLS settings for pgo-apiserver
The default minimum TLS was 1.1, which is deprecated. Switching the minimum to
1.2 should have no adverse affects on clients, which will be using 1.2/1.3
already.

Additionally, in this commit the ciphersuites used for TLS 1.2 are restricted
to a known good subset, based around the following resources:

- https://wiki.openssl.org/index.php/FIPS_mode_and_TLS#TLS_1.2
- https://github.com/ssllabs/research/wiki/SSL-and-TLS-Deployment-Best-Practices
- https://blog.qualys.com/product-tech/2019/04/22/zombie-poodle-and-goldendoodle-vulnerabilities

This enhances the baseline TLS configuration of pgo-apiserver, with a view to
making it possible to configure this during PGO installation or via configmap
in the future.

Signed-off-by: Steve Kerrison <steve@usec.io>