Allow for startup to proceed even if status subresource is missing
We have a guard to only allow a cluster to be started up if we know
that it is in a shut down state. However, there appears to be cases
independentof the Operator where the pgclusters.crunchydata.com
"status" subresource is missing, which would cause that guard to be
too aggressive.

This change allows for the condition that there is no state registered
in the status subresource, which allows for the start up to proceed
regardless.

Issue: [ch10811]