Add OwnerReference to root cert secret
This commit adds one or more postgresclusters as owner references
to the root cert secret in the same namespace. This allows for the
create / garbage collection of the root cert secret to coincide with
the creation / deletion of the postgresclusters in the same namespace.
Please note that the "Controller" value is set to false in all cases
to allow for more than one OwnerReference at a time.