Maintain YAML & JSON Ordering for DCS Config
Testing has shown that with versions of PostgreSQL earlier than v13,
changes to the ordering of the various Patroni and PostgreSQL settings
stored within a cluster's DCS (as stored in the "config" annotation of
the "<cluster-name>-config" ConfigMap) can cause a replica to
automatically restart.  And being that the go standard library does not
guarantee ordering when encoding/decoding JSON, the DCS configuration
can be re-ordered when the "<cluser-name>-pgha-config" ConfigMap is
updated, and the operator subsequently syncs its contents with the
"<cluster-name>-config" ConfigMap.  This, in turn, can lead to the
automatic restart of a replica following a configuration change in the
"<cluser-name>-pgha-config", since the settings might be patched in a
different order than initially stored.

While in general order for JSON typically should not matter, this
commit ensures order is maintained when marshalling and unmarshalling
the DCS configuration found within both the "<cluster-name>-config" and
"<cluster-name>-pgha-config" ConfigMaps into both YAML and JSON.  With
this change,a replica will no longer automatically following a
configuration change, but will rather show a pending restart as
expected.