Update container spec to run uid script
To support running in Openshift containers created by the operator need to
support using a random UID that is defined by Openshift. This change updates the
container spec for the database pod to use an entrypoint script
(uid_postgres.sh) as the container command (Kubernetes equivalent to Docker
entrypoint). To support this change Patroni is now started using container args
(Kubernetes equivalent to Docker command).