PR review updates
Per the discussion on the PR, this removes the root cert from being
bundled with the leaf certificate and key used by Patroni's TLS
identification "patroni.crt-combined" file. It also clarifies the
usage of "openssl verify -CAfile" and cleans up some of the test
implementations and corrects/clarifies several comments.