Allow for cluster-wide PVC resizing
This adds the ability to resize the PVC that contains the data
directory of a PostgreSQL instance and applies the changes
cluster-wide. This uses the standard Kubernetes formula for applying
a PVC change, i.e:

- Each PVC needs to be unmounted from the Pod for a PVC change to
take effect. The Operator handles this by scaling down/up each
Deployment
- The storage class assigned to the PVC must support resizing
- The new size of the PVC must be larger than the old size

Changes are applied using a rolling update to minimize downtime.

If a specific Postgres instance (pgreplica) has a size that
differs from the cluster size, it is skipped, unless the new
cluster size is larger than the instance size.

Issue: [ch8477]