Adds nss_wrapper Support for RORFS in OpenShift
Adds support for the nss_wrapper as implemented within the Crunchy
Container Suite, specifically as needed for compatibility with the
'restricted' SCC in OpenShift (and therefore the random UID's that are
assigned to containers by OpenShift).  This is done in accordance with
RedHat guidance (see https://access.redhat.com/articles/4859371),
and ensures a 'readOnlyRootFileSystem' can be enabled when the
PostgreSQL Operator is deployed in an OpenShift environment.

To support the nss_wrapper solution, applicable PostgreSQL and
pgBackRest Deployment templates have been updated with the proper
nss_wrapper environment variables.  Additionally, the default SSHD
configuration file has also been updated to ensure the proper
nss_wrapper environment variables are set when pgBackRest executes
various commands over SSH.