Reconcile a simple PgBouncer Deployment
It is now possible to describe a PgBouncer instance that will proxy
connections to the primary instance of a PostgresCluster. A PgBouncer
user, schema, and function to access passwords is installed into the
PostgreSQL database cluster so that user passwords configured there may
also be used when connecting through PgBouncer.

This functionality is entirely optional in the PostgresCluster
specification. When omitted, no PgBouncer objects are created in
Kubernetes nor PostgreSQL. The specification can be added or removed at
any time and the objects will be created and deleted appropriately.

To avoid storing the plaintext password in a Secret, the PgBouncer user
authenticates to PostgreSQL using an MD5 password. This does not affect
other users which may have SCRAM passwords in PostgreSQL.

The PgBouncer admin console would allow a person or system to manipulate
and monitor connections, but no users are authorized to access it at
this time. More advanced authentication and management will happen in a
later commit.

Issue: [ch10774]
Issue: [ch10787]
Issue: [ch10789]