Allow for a S3 bucket name to be edited after cluster creation
This allows for the name of a S3 bucket to be edited after a
cluster is created, i.e. through the custom resource. Should this
action be taken, the following occurs:

- The pgBackRest report Pod is redeployed
- The PostgreSQL cluster is redeployed using a rolling update
- Once the above two actions complete, a new stanza is created and
an initial backup in that stanza is created.

A create stanza must be attempted as the new bucket may be
entirely empty.

Issue: [ch11061]