Allows Restores Across Namespaces (#2358)
When creating a new PostgreSQL cluster via 'pgbackrest restore' using
the pgBackRest backups from another current or former PostgreSQL
cluster, it is now possible restore from the backups of a cluster in
different namespace (i.e. different than the cluster being created).

In support of this functionality, a copy of the pgBackRest Secret
corresponding to the repository containing the backups being utilized
to bootstrap the cluster is now made specifically for the bootstrap Job
(and then removed when bootstrapping is complete).  This ensures the
new cluster always has the information required to connect to the
repository being utilized to bootstrap the cluster, regardless of what
namespace it is in.  Additionally, DNS for the pgBackRest repository
service is now always used when performing the restore to ensure
proper connectivity across namespaces if/when needed.

And finally, a  '--restore-from-namespace' option has been added to the
pgo client, which corresponds to a 'spec.pgdatasource.namespace' field
in the Pgcluster spec, and allows users to specify the namespace of the
cluster specified for bootstrapping the new cluster.

Issue: [ch9969]
Issue: #2062