Allow the pgBackRest PVC to be resized
The `size` attribute of the `BackrestStorage` section of the
`pgclusters.crunchydata.com` can be modified to resize the PVC
for a pgBackRest cluster. Similar to a cluster PVC resize, such
a change needs to follow the same rules:

- Each PVC needs to be unmounted from the Pod for a PVC change to
take effect.
- The storage class assigned to the PVC must support resizing
- The new size of the PVC must be larger than the old size

Additionally, this only affects the local storage pgBackRest
PVC and not an external storage systems (e.g. S3).

As there is only a single pgBackRest Deployment, this triggers a
scale down/up operation of that Deployment to make the change.

Issue: [ch8477]
Issue: #1626
Issue: #2332