Allow a single instance/replica to have its PVC resized
This allows for a single instance/replica to have its individual PVC
resized.  Similar to the other PVC operations, it must meet
the following requirements:

- Each PVC needs to be unmounted from the Pod for a PVC change to
take effect.
- The storage class assigned to the PVC must support resizing
- The new size of the PVC must be larger than the old size

If a cluster-wide PVC size change is subsequently initiated, the
Operator will ignore the change for the specific instance, unless the
cluster change has a larger value than the instance value.

Issue: [ch8477]