Enable mTLS  for Replication & Rewind Accounts
This commit updates the implementation of the replication
and rewind system accounts to leverage the TLS infrastructure
implemented within the Operator. The Operator now uses a more
secure credential in the form of a client certificates in place
of the existing MD5 usernames and passwords for authenticating
into the PostgreSQL database. These certificates can either be
provided during deployment or autogenerated by the Operator
automatically.

The required Patroni settings for the mutual authentication of
the replication and rewind system accounts is generated so that
the appropriate certificates are used when connecting to the
database. Additionally, the pg_hba.conf file is configured via
Patroni to ensure the proper (and secure) authentication of these
accounts. Please note that the sslmode setting is currently
'verify-ca' due to a known issue
(https://github.com/zalando/patroni/issues/1884).