Allow PgBouncer to be configured in great detail
PgBouncer is able to perform a variety of functions depending on how it
is configured. Most choices involve trade-offs that are best left to the
user. The "spec.proxy.pgBouncer.config" section can be used to adjust
any PgBouncer setting.

To apply those settings to PgBouncer with minimal downtime, a sidecar
now runs in the PgBouncer pod watching for any changes to configuration
files. When it detects a change, it signals PgBouncer to reload.

Issue: [ch10776]