Create a ServiceAccount for each PostgresCluster
Patroni calls to the Kubernetes API to elect a write-leader and maintain
consistent configuration across instances. To do this, it must identify
itself using a ServiceAccount in its same namespace.

The PostgresCluster controller now creates that ServiceAccount and
grants it exactly the permissions necessary for that cluster
specification.

Issue: [ch10729]