Add ability to set Postgres password type at cluster creation
Specifically, this allows one to use "scram-sha-256" formatted
passwords when a PostgreSQL cluster is created.

This adds a new attribute to the `pgclusters.crunchydata.com` CRD
called "passwordType", which stores the default value of the password
hashing mechanism that the Postgres cluster should use.

This also adds the "--password-type" flag to "pgo create cluster",
which accepts the same values as the user-oriented commands, i.e.
"scram-sha-256" and "md5".

If a password type is not provided when a new user is created, the
value from the custom resource is used.

Issue: [ch11049]