Cleans pgBackRest Repo on In-Place Restore
The pgBackRest repository is now cleaned up following an in-place
restore and then recreated during cluster initialization.  This
ensures the proper initialization logic is executed when the
pgBackRest repository is scaled up to one during initialization.
Specifically, this triggers the deployment and initialization of the
Primary Pod for the cluster.