Cleanup PostgreSQL Instances (#51)
* Cleanup PostgreSQL Instances

This is the first changeset to allow the operator to clean up instances
that have been removed from the Postgrescluster spec. After the cluster
has been updated to include new instances (scaled up), scale down logic
will be called to remove instances that meet one of the following
conditions:

- Instances where the instance set has been removed
- Instances included in an instance set that was updated to have fewer
replicas

In either of these cases the primary pod can be deleted forcing the
cluster to failover. If the instance set including the primary is
deleted a failover will occur. If the instance set including the primary
is updated to have 0 replicas the primary will be deleted. The primary
will only be deleted if one of these two conditions are met and it will
always be the last instance to be deleted.

When an instance is deleted its statefulsets, secrets, configmaps, and
PVCs are deleted using the cluster and instance labels. Any secret or
configmap that gets added in the future, and includes these labels, will
be deleted as part of the scale down.

The reconcileInstanceSet() function has been refactored. This is now
scaleUpInstances() and is called using a new function
reconcileInstanceSets(). The new function has the same return value as
the original function. ScaleDownInstances() returns an error if one
occurred during the scale down process.

* Update internal/controller/postgrescluster/instance.go

squash - comment update

* comment updates from pr

* comment update from pr

* use controlled delete function

* update test names

* updates for tests