Enables Automatic pgBackRest Stanza Creation
Stanzas are now automatically created for any/all pgBackRest
repositories defined in the PostgresCluster spec.  Additionally,
stanzas are created for any new repositories added throughout
the lifetime of a PostgreSQL cluster.

To properly detect changes to pgBackRest repository configuration
and execute stanza create commands accordingly, all repositories must
now be defined within the PostgresCluster spec (i.e. repositories
should not be defined solely via custom ConfigMaps/Secrets if automated
pgBackRest orchestration, e.g. automatic stanza creation, is desired).
In support of this requirement, the PostgresCluster spec has been
updated to not only allow repository volumes (PersistentVolumeClaims)
to be defined, but to also allow external pgBackRest repositories to be
defined (e.g. Azure, GCS, and/or S3 repositories).

As stanza creation succeeds or fails for the various repositories
defined for in the PostgresCluster spec, the pgBackRest repository
status is updated accordingly to indicate whether or not a successful
stanza creation has been completed for each repository.  A hash of the
various configuration settings for the external repositories as defined
in the spec is utilized to track and detect updates to external
repository configuration, therefore allowing for the re-execution of
the stanza create command when changes to the external configuration
settings are detected (e.g. when bucket names, endpoints, regions,
storage containers, etc. are modified).

Additionally, events are generated upon successful stanza creation,
while warning events are generated for failed stanza creation attempts.
This ensures full end-user visibility into stanza creation activities
and issues.

And finally, 'archive_mode' is now set to 'on' by default as required
by pgBackRest, and the proper pgBackRest 'archive_command' is now also
configured by default.  This ensures WAL can be pushed by PostgreSQL
and backups can be taken following successful stanza creation.

It should also be noted that a fix is included to address a race
condition between a PG instance going ready, and Patroni updating the
config Endpoints with the system identifier for the newly bootstrapped
cluster.  Specifically, if a ready instance is detected but the system
identifier cannot be found in the Endpoint, the Operator will now
simply re-queue and try again until the expected identifier is found.

[ch9481]