Adds nss_wrapper Support
Updates the Pod template for all PostgreSQL instances and the
pgBackRest dedicated repository host (if enabled) to include the proper
environment variables for the nss_wrapper, which is now now implemented
in the "postgres-ha" and "pgbackrest" containers for compatibility with
OpenShift (see https://access.redhat.com/articles/4859371).  The
environment variables specifically ensure pgBackRest commands can be
run within the "database" and "pgbackrest" containers without first
requiring environment variables to be set, the use of wrapper scripts,
or other workarounds.

Additionally, a ForceCommand setting is now been included in the
default 'sshd_config' file that is created when a pgBackRest repo host
(and therefore SSHD) is enabled for a specific PostgresCluster.  This
ensures the proper nss_wrapper environment variables are present in
the environment when pgBackRest utilizes SSH to communicate with a
specific host.

And finally, an init container has been added to all PostgreSQL
instance Pods, as well as the dedicated repo host Pod (if enabled),
as needed to pre-configure the nss_wrapper environment before the
actual containers for these Pods run.  This prevents the need to
include nss_wrapper logic in the commands configured for all "database"
and "pgbackrest" containers, while also preventing nss_wrapper errors
from being displayed early during container initialization as a result
of the nss_wrapper environment variables being present before
nss_wrapper setup logic has had a chance to execute.

[ch11282]