Rolling update for changes to PostgresInstanceSetSpec
We deploy PostgreSQL instances as individual StatefulSets so that:

 - a Kubernetes controller keeps the Pods running,
 - each Pod has a stable DNS name, and
 - each Pod can be stopped and started independently

This commit leverages that last property to orchestrate rolling updates
of instance specifications while considering each Pod's availability and
role as primary or replica of the PostresCluster.

To that end, a new Instance type combines the PostgresInstanceSetSpec,
StatefulSet, and Pod objects of each logical instance to make it easy to
consider them together. It has methods like IsPrimary and IsAvailable
that can be useful outside of the rolling update logic.

A new "instances" section is added to PostgresClusterStatus that shows
the progress of rolling updates. Its "replicas", "readyReplicas", and
"updatedReplicas" fields have the same semantics as fields with those
names in appsv1.DeploymentStatus and appsv1.StatefulSetStatus.

Issue: [ch10193]
Issue: [ch11299]