Switchover the primary during a rolling update
When stopping due to SIGTERM, Patroni performs a "fast" shutdown of
PostgreSQL. A failover will occur when that shutdown takes longer than
Patroni's configured "ttl" on the primary.

Instead, perform a switchover on the primary so that Patroni stops
PostgreSQL using what it calls "graceful" mode. It performs an immediate
checkpoint in the background then performs a "fast" shutdown when that
completes. The recent checkpoint makes this shutdown swift and safe.

Unfortunately, the "patronictl" interface for triggering switchover has
a number of deficiencies:

 - It exits zero even when the switchover does not complete.
 - It waits up to 2 × "loop_wait" for the election to complete.
 - It rejects scheduled switchover that it believes is in the past,
   making it unpredictable when throttled or due to clock skew.

Issue: [ch10193]