Custom Labels
This update allows, users to define custom labels for any/all resources
created by the PostgreSQL Operator. Updating the PostgresCluster Spec
with custom labels in the following fields will be added to resources
when they are reconciled:
- `spec.metadata.labels` (cluster)
- `spec.archive.pgbackrest.metadata.labels` (pgbackrest)
- `spec.proxy.pgbouncer.metadata.labels` (pgbouncer)
- `spec.instances[i].metadata.labels` (instance)

The labels that get added to a resource are a combination of cluster
labels that are added globally, instance, pgbackrest, or pgbouncer
labels that are added to their corresponding resources, and the
non-user labels that the PostgreSQL Operator adds normally. When these
labels are merged non-user labels will always win, then specific
resource labels, then global labels.

End-to-end style tests have been added to ensure that when a cluster is
created and reconciled with custom labels, every resource that it creates
is labeled. The tests use resource labels to determine what resources
should be labeled. For example, PGBackRest tests list resources with the
`postgres-operator.crunchydata.com/pgbackrest` label and make sure that
it also has any label defined in `spec.archive.pgbackrest.metadata.labels`.