pgBackRest cronjob schedules (#74)
* pgBackRest cronjob schedules

This commit adds the initial ability to add a cronjob for
performing pgBackRest backups on a given schedule to the
postgrescluster CRD. It allows for Kubernetes CronJobs to be
created for each pgBackRest repo for each of the three backup
types (full, differential and incremental). If at a later point
the schedules are removed, the relevant CronJob, along with the
associated Job(s) and Pod(s) are also removed.

Currently, the job simply prints a message at the set time
interval, but will allow for integration with the related
pgBackRest backup feature.

* Updates from PR Review

This commit contains the updates requested during
the PR review process.