Enables pgBackRest Replica Creation
pgBackRest is now utilized as the primary mechanism for creating new
replicas in a PostgreSQL cluster.  Specifically, once a cluster has
been bootstrapped and a stanza has been created for the various repos
configured in the PostgresCluster spec, a backup will automatically
be taken for the repo defined at index 0 per the "repos" array within
the spec.  And following the successful completion of this backup, a
"pgbackrest" replica creation method will be enabled via Patroni
configuration, after which all newly created replicas will be created
via a restore from this specific  repository (i.e. the repo defined at
index 0).

With this design, any replicas that are created very early within
the PostgreSQL cluster lifecyle will still likely be created using
pg_basebackup.  This is because pg_basebackup (the default replica
creation implemented by Patroni) will remain in place as the default
replica creation method until a backup has completed successfully as
described above.

Additionally (and as mentioned above), the repo defined at index 0 in
the postgrescluster.spec.archive.pgbackrest.repos array now plays a
special role in the cluster initialization process.  Specifically, this
repo is designated as the "replica creation repo", and will therefore
be utilized for bootstrapping (or re-initializing) replicas within the
cluster (specifically once its stanza and initial backup have both
completed successfully).  This means that as the repo at index 0
changes throughout the lifetime of a cluster (e.g. because the repos in
the array are reorganized or a configuration change is made) then a
backup will be taken (and then Patroni reconfigured as needed) to
ensure proper (and continued) pgBackRest replica creation.

[ch11104]