Handle errors expected from PVC writes
A PersistentVolumeClaim's storage request cannot change until after it
binds to a PersistentVolume. Changing the request in the cluster spec
soon after scaling up can trigger an error. When this happens, emit an
event rather than fail to reconcile.

Most other changes to a PVC spec are also disallowed, but preventing
these requires a validating webhook. Instead, set a condition and emit
events for these as well.

Issue: [ch11119]