Replication TLS Certificate and Key Copy Sidecar
This commit adds a sidecar container that copies the mounted replication
client TLS certificate, key and CA certificate files from the mounted
projected volume directory (/pgconf/tls/replication) to the /tmp/replication
directory in order to set proper file permissions. This function involves a
continual loop that checks for changes to the relevant mounted directory. If
a change is detected, the files are copied over and proper permissions are set.
Finally, a Patroni reload is executed so the new configuration is used. As
during initialization, this is required because the group permission settings
applied via the defaultMode option are not honored as expected, resulting in
incorrect group read permissions.

Note that this commit also updates the name of the existing 'CopyReplicationTLS'
to 'InitCopyReplicationTLS' to better reflect the current usage.