Custom Annotations
This update allows users to define custom annotations for any/all
resources created and owned by the PostgreSQL Operator. Updating the
PostgresCluster Spec with custom annotations in the following fields
will be added to resources when they are reconciled:

- spec.metadata.annotations (cluster)
- spec.archive.pgbackrest.metadata.annotations (pgbackrest)
- spec.proxy.pgbouncer.metadata.annotations (pgbouncer)
- spec.instances[i].metadata.annotations (instance)

Annotations added to a resource are a combination of annotations defined in
the metadata field of clusters and metadata fields of instance specs,
pgbackrest repos, or pgbouncer specs define in the PostgresCluster spec.

When annotations are merged, non-custom annotations will always win,
then specific resource annotations (for instance, pgbackrest, or
pgbouncer), then annotations defined globally for the cluster.