Consider posix,s3 & posix,gcs During Upgrade
The Operator upgrade process now properly considers the presence of
both "posix" and "s3" storage, or both "posix" and "gcs" storage, for
the backrestStorageTypes field within the pgcluster spec. Specifically,
when updating the "archive_command" during a pgcluster upgrade
(specifically due to a directory change in v4.6 impacting the directory
utilized for the "archive_command"), a check is now performed to
determine if both "posix" and "S3" storage is enabled, or if both
"posix" and "GCS" storage is enabled, and the set the proper
"archive_command" is then set accordingly.