Rearrange PostgreSQL data volume setup
This moves file paths and some other constants from the naming package
to the postgres package. Refer to the PVC as the "PostgreSQL data
volume" to distinguish it from the $PGDATA environment variable which
does a few different things. Also simplify mounting that volume by
initializing the PostgreSQL containers together.