Update addNSSWrapper function to take image
The addNSSWrapper function now takes the image that should be used to
define the init container as a function. When adding the NSS wrapper
init container to a template spec you should provide the image that
should be used. Currently this is either the spec.Image for an instance
or spec.Archive.PGBackRest.Image for a repo host.

Issue: [ch11433]